AdminのFeatureテスト実装

+++ src/mealbox/app/Http/Controllers/Admin/AdminController.php
@@ -30,7 +30,7 @@ class AdminController extends Controller
         $user_id = $request->input('user_id');
         $password = $request->input('password');
 
-        if ($user_id === config('admin.admin_id') && $password === config('admin.admin_pass')) {
+        if ($user_id == config('admin.admin_id') && $password == config('admin.admin_pass')) {
             // ログインの成功
             $request->session()->put("admin_auth", true);
             return redirect()->route("admin.top");
@@ -138,6 +138,10 @@ class AdminController extends Controller
      */
     function adminOrderDelete($id)
     {
+        if (empty($id)) {
+            return redirect()->route('admin.order_list');
+        }
+
         try {
             Order::destroy($id);
         } catch(\Throwable $e) {
